+++ client/src/layout/app/AppSidebar.vue
@@ -3,7 +3,7 @@ import AppMenu from './AppMenu.vue';
 </script>
 
 <template>
-    <app-menu></app-menu>
+  <app-menu></app-menu>
 </template>
 
 <style lang="scss" scoped></style>
