+++ client/src/layout/app/AppTopbar.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { ref, computed, onMounted, onBeforeUnmount } from 'vue';
-import { useLayout } from '@/layout/composables/layout';
+import { useLayout } from '@/layout/app/composables/layout.js';
 import { useRouter } from 'vue-router';
 
 const { layoutConfig, onMenuToggle } = useLayout();
