+++ client/src/main.js
@@ -17,19 +17,23 @@ app.use(createPinia());
 //#endregion
 
 //#region tanstack query
+
 import { VueQueryPlugin } from '@tanstack/vue-query';
 app.use(VueQueryPlugin);
+
 //#endregion
 
 //#region primevue
+
 import PrimeVue from 'primevue/config';
 app.use(PrimeVue, {
   // ripple: true,
 });
 
-// import 'primeicons/primeicons.css';
+import '@/styles/prime.scss';
 import 'primevue/resources/themes/lara-light-green/theme.css';
-import '@/styles/styles.scss';
+
+//#region components
 
 import Menubar from 'primevue/menubar';
 // eslint-disable-next-line vue/multi-word-component-names
@@ -43,6 +47,9 @@ app.component('Menu', Menu);
 import Button from 'primevue/button';
 // eslint-disable-next-line vue/multi-word-component-names,vue/no-reserved-component-names
 app.component('Button', Button);
+
+//#endregion
+
 //#endregion
 
 //#region uid
