+++ client/src/router/index.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router';
 import { routes } from '@/router/routes';
 import authService from '@/services/auth/auth.service';
 import { useAuthStore } from '@/stores/auth';
-import AppLayout from '@/layout/AppLayout.vue';
+import AppLayout from '@/layout/app/AppLayout.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -28,19 +28,19 @@ const router = createRouter({
             {
               name: routes.chores.children.overview.name,
               path: routes.chores.children.overview.path,
-              component: () => import('@/pages/chores/ChoresOverview.vue'),
+              component: () => import('@/pages/app/chores/ChoresOverview.vue'),
             },
             {
               name: routes.chores.children.detail.name,
               path: routes.chores.children.detail.path,
-              component: () => import('@/pages/chores/ChoreDetail.vue'),
+              component: () => import('@/pages/app/chores/ChoreDetail.vue'),
             },
           ],
         },
         {
           name: routes.tags.name,
           path: routes.tags.path,
-          component: () => import('@/pages/tags/ManageTags.vue'),
+          component: () => import('@/pages/app/tags/ManageTags.vue'),
         },
       ],
     },
