--- client/src/styles/app.scss
@@ -1,3 +1,0 @@
-#app {
-  //padding: 1rem;
-}
