+++ client/src/styles/layout/_content.scss
@@ -1,12 +1,12 @@
 .layout-main-container {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-    justify-content: space-between;
-    padding: 7rem 2rem 2rem 4rem;
-    transition: margin-left $transitionDuration;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  justify-content: space-between;
+  padding: 7rem 2rem 2rem 4rem;
+  transition: margin-left $transitionDuration;
 }
 
 .layout-main {
-    flex: 1 1 auto;
+  flex: 1 1 auto;
 }
