+++ client/src/styles/layout/layout.scss
@@ -1,12 +1,12 @@
 @import './_variables';
-@import "./_mixins";
-@import "./_preloading";
-@import "./_main";
-@import "./_topbar";
-@import "./_menu";
-@import "./_config";
-@import "./_content";
-@import "./_footer";
-@import "./_responsive";
-@import "./_utils";
-@import "./_typography";
+@import './_mixins';
+@import './_preloading';
+@import './_main';
+@import './_topbar';
+@import './_menu';
+@import './_config';
+@import './_content';
+@import './_footer';
+@import './_responsive';
+@import './_utils';
+@import './_typography';
