+++ client/src/styles/prime.scss
@@ -1,0 +1,3 @@
+$gutter: 1rem; //for primeflex grid system
+@import '@/styles/layout/layout.scss';
+@import 'primeicons/primeicons.css';
