--- client/src/styles/styles.scss
@@ -1,12 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-
-$gutter: 1rem; //for primeflex grid system
-@import '@/styles/layout/layout.scss';
-
-/* PrimeVue */
-// @import 'primevue/resources/primevue.min.css'; // version deplicated. You must open the command line if you are using < 3.29.00
-//@import 'primeflex/primeflex.scss';
-@import 'primeicons/primeicons.css';
-
-/* Demos */
-//@import '@/styles/demo/demo.scss';
